Fix TFS branch+edit operations.
The differ for branch+edits wasn't correctly including the original filename or
revision in the diff. The result of this was that files which were copied only
would show up properly, though the "Copied from ..." line would be wrong. Files
that were copied and then edited would fail to patch, since the original
version wasn't correctly found.

This change fixes it up so we set the `oldFilename` and `oldVersion` correctly
for both renames and copies.

Testing done:
Posted a change that involved a branch+edit of a file. Before this change, I
saw a patch failure in the diff viewer. After, I see the diff I expect.

Reviewed at https://reviews.reviewboard.org/r/8666/