Catch VersionControlException when doing diffs.
A number of things in the TFS SDK can throw VersionControlException, which is
apparently an unchecked exception, and therefore wasn't included in any of the
docs for the affected methods. This was showing up in a customer's environment
where a timeout when fetching data from the server wasn't causing "rbt post" to
fail, but the resulting broken diff would fail to apply.

This change adds the relevant type to the catch, and fixes up the
alphabetization on the imports. This also fixes a place where I forgot to
update the version number of the TFS SDK.

Testing done:
Built the adapter and pushed to AWS. Smoke tested "rbt diff" and "rbt post".

Reviewed at https://reviews.reviewboard.org/r/8671/